test: move to seperate file
